requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
cached-property==1.5.2
certifi==2024.2.2
distro==1.9.0
dj-database-url==2.1.0
Django==3.2.24
django-heroku==0.3.1
exceptiongroup==1.2.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
importlib-metadata==6.7.0
openai==1.13.3
packaging==24.0
psycopg2==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==0.21.1
pytz==2024.1
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.7.1
whitenoise==6.5.0
zipp==3.15.0
